refactor(cli): extract shared help/exit module src/cli/help.ts
Relocates the help/version report printers, the TUI-aware exit helpers,
the promptYesNo/confirmPrompt readline core, and BINARY_ONLY_COMMANDS +
USAGE out of cli/bakin.ts verbatim (dynamic-import specifiers adjusted
for the new location). Behavior-preserving; shared by the upcoming
per-scope command modules.
